requirements.txt

pandas
numpy
sentence-transformers
faiss-cpu
scikit-learn
hdbscan
umap-learn
xgboost
fastapi
uvicorn
google-generativeai
python-dotenv
